Partially reorganize Lua API classes to more logical form
