+++ luascript/api/adventure/StackInstance.cpp
@@ -11,10 +11,10 @@
 
 #include "StackInstance.h"
 
-#include "Registry.h"
+#include "../Registry.h"
 
-#include "../LuaStack.h"
-#include "../LuaCallWrapper.h"
+#include "../../LuaStack.h"
+#include "../../LuaCallWrapper.h"
 #include <vcmi/Creature.h>
 
 VCMI_LIB_NAMESPACE_BEGIN
