+++ luascript/api/library/HeroClass.cpp
@@ -11,17 +11,16 @@
 
 #include "HeroClass.h"
 
-#include "Registry.h"
+#include "../Registry.h"
 
-#include "../LuaStack.h"
-#include "../LuaCallWrapper.h"
+#include "../../LuaStack.h"
+#include "../../LuaCallWrapper.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
 
-namespace scripting
+namespace scripting::api::library
 {
-namespace api
-{
+
 const std::vector<HeroClassProxy::CustomRegType> HeroClassProxy::REGISTER_CUSTOM =
 {
 	{"getIconIndex", LuaMethodWrapper<HeroClass, decltype(&Entity::getIconIndex), &Entity::getIconIndex>::invoke, false},
@@ -30,7 +29,6 @@ const std::vector<HeroClassProxy::CustomRegType> HeroClassProxy::REGISTER_CUSTOM
 	{"getName", LuaMethodWrapper<HeroClass, decltype(&Entity::getNameTranslated), &Entity::getNameTranslated>::invoke, false},
 };
 
-}
 }
 
 VCMI_LIB_NAMESPACE_END
