+++ luascript/api/library/Skill.cpp
@@ -11,17 +11,16 @@
 
 #include "Skill.h"
 
-#include "Registry.h"
+#include "../Registry.h"
 
-#include "../LuaStack.h"
-#include "../LuaCallWrapper.h"
+#include "../../LuaStack.h"
+#include "../../LuaCallWrapper.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
 
-namespace scripting
+namespace scripting::api::library
 {
-namespace api
-{
+
 const std::vector<SkillProxy::CustomRegType> SkillProxy::REGISTER_CUSTOM =
 {
 	{"getIconIndex", LuaMethodWrapper<Skill, decltype(&Entity::getIconIndex), &Entity::getIconIndex>::invoke, false},
@@ -30,7 +29,6 @@ const std::vector<SkillProxy::CustomRegType> SkillProxy::REGISTER_CUSTOM =
 	{"getName", LuaMethodWrapper<Skill, decltype(&Entity::getNameTranslated), &Entity::getNameTranslated>::invoke, false},
 };
 
-}
 }
 
 VCMI_LIB_NAMESPACE_END
